Array based problem
Write a program to perform the following operations using Java.
1. Read the count of elements from user to create the array.
2. Inttialize the array using user input. 
3. Create the functions for the following and call them using a menu based program.
a. Display the array
b. Reverse the array
c. Replace the nth element from the array and print
d. Find the addition of all elements of the array
e. Find the maximum number from the array
f. Print the average of all the elements from the array.
g. Display the even and odd number from the array separately